feat(post): Likes and comments implemented

+++ src/main/java/br/com/orangex/api/service/PostService.java
@@ -1,16 +1,21 @@
 package br.com.orangex.api.service;
 
+import br.com.orangex.api.dto.CreateCommentDTO;
 import br.com.orangex.api.dto.CreatePostDTO;
 import br.com.orangex.api.dto.UpdatePostDTO;
 import br.com.orangex.api.exception.AuthException;
 import br.com.orangex.api.exception.NotFoundException;
+import br.com.orangex.api.model.Comment;
 import br.com.orangex.api.model.Post;
 import br.com.orangex.api.model.User;
+import br.com.orangex.api.repository.CommentRepository;
 import br.com.orangex.api.repository.PostRepository;
 import br.com.orangex.api.repository.UserRepository;
 import br.com.orangex.api.util.CurrentUser;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
@@ -26,6 +31,9 @@ public class PostService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private CommentRepository commentRepository;
+
     public Post create(CreatePostDTO postDTO){
 
         Post newPost = new Post();
@@ -36,7 +44,6 @@ public class PostService {
         if(newPost.getAuthor().username().equals(currentUser.getUsername()) && newPost.getAuthor().name().equals(currentUser.getName())){
             User user = userRepository.findById(currentUser.getId()).get();
 
-            newPost.setLikes(0);
             newPost = postRepository.save(newPost);
 
             user.getPosts().add(newPost);
@@ -47,6 +54,10 @@ public class PostService {
             throw new AuthException("An authentication error occurred...");
         }
 
+    }
+
+    public Page<Post> getAll(Pageable pageable){
+        return postRepository.findAll(pageable);
     }
 
     public Post getById(String id){
@@ -78,9 +89,46 @@ public class PostService {
         Optional<Post> post = postRepository.findById(id);
         if(post.isEmpty()) throw new NotFoundException("Post", id);
 
-        if(!post.get().getAuthor().username().equals(CurrentUser.getCurrentAuthenticatedUser().getUsername())) throw new AuthException("An authentication exception ocurred...");
+        if(!post.get().getAuthor().username().equals(CurrentUser.getCurrentAuthenticatedUser().getUsername()) || !post.get().getAuthor().name().equals(CurrentUser.getCurrentAuthenticatedUser().getName())) throw new AuthException("An authentication exception ocurred...");
 
         postRepository.deleteById(id);
+
+    }
+
+    public Integer like(String id){
+
+        Post post = getById(id);
+
+        boolean alreadyLiked = post.getLikes().contains(CurrentUser.getCurrentAuthenticatedUser().getUsername());
+
+        if(alreadyLiked){
+            post.getLikes().remove(CurrentUser.getCurrentAuthenticatedUser().getUsername());
+        }else{
+            post.getLikes().add(CurrentUser.getCurrentAuthenticatedUser().getUsername());
+        }
+
+        postRepository.save(post);
+
+        return post.getLikes().size();
+
+    }
+
+    public Post comment(String id, CreateCommentDTO commentDTO){
+
+        Post post = getById(id);
+
+        if(!commentDTO.author().username().equals(CurrentUser.getCurrentAuthenticatedUser().getUsername()) || !commentDTO.author().name().equals(CurrentUser.getCurrentAuthenticatedUser().getName())) throw new AuthException("An authentication error occured...");
+
+        Comment comment = new Comment();
+
+        BeanUtils.copyProperties(commentDTO, comment);
+
+        comment = commentRepository.save(comment);
+
+        post.getComments().add(comment);
+        postRepository.save(post);
+
+        return post;
 
     }
